go-server/pyproject.toml
[project]
name = "websocket-test"
version = "0.1.0"
description = "WebSocket testing for Go server"
authors = [
    {name = "Hackathon"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "websockets (>=15.0.1,<16.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
